+++ sdk/common/src/main/java/hr/avrbanac/skg/sdk/common/KafkaDtoStatus.java
@@ -1,0 +1,10 @@
+package hr.avrbanac.skg.sdk.common;
+
+public enum KafkaDtoStatus {
+    INIT,
+    OK,
+    ERROR,
+    FALLBACK,
+    RESERVED,
+    UNKNOWN
+}