+++ sdk/common/src/main/java/hr/avrbanac/skg/sdk/common/KafkaDtoType.java
@@ -1,0 +1,11 @@
+package hr.avrbanac.skg.sdk.common;
+
+public enum KafkaDtoType {
+    AUDIT,
+    CONTROL,
+    GENERAL,
+    PERIODIC,
+    TEST,
+    EVENT,
+    SPECIAL
+}